progress: split tracker.go into tracker_sort.go
